feat: add image upload functionality for products
Introduce the ability to upload images for products in the admin panel. This includes adding a new field `ImageURL` to the `Product` model, creating a new handler for image uploads, and configuring static file serving for uploaded images. The `github.com/google/uuid` package is added to generate unique filenames for uploaded images. The `.gitignore` file is updated to exclude the `uploads` directory from version control.